@@ -1,10 +1,10 @@
 name: month_picker_dialog
 description: Internationalized dialog for picking a single month from an infinite list of years.
-version: 0.3.3
+version: 1.0.0
 homepage: https://github.com/hmkrivoj/month_picker_dialog
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -12,8 +12,8 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   intl: ">=0.1.0<2.0.0"
-  flutter_cupertino_localizations: ^1.0.1
-  rxdart: ^0.24.0
+  #flutter_cupertino_localizations: ^1.0.1
+  rxdart: ^0.26.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
