@@ -8,17 +8,17 @@ import 'locale_utils.dart';
 
 class YearSelector extends StatefulWidget {
   final ValueChanged<int> onYearSelected;
-  final DateTime initialDate, firstDate, lastDate;
+  final DateTime? initialDate, firstDate, lastDate;
   final PublishSubject<UpDownPageLimit> upDownPageLimitPublishSubject;
   final PublishSubject<UpDownButtonEnableState>
       upDownButtonEnableStatePublishSubject;
-  final Locale locale;
+  final Locale? locale;
   const YearSelector({
-    Key key,
-    @required this.initialDate,
-    @required this.onYearSelected,
-    @required this.upDownPageLimitPublishSubject,
-    @required this.upDownButtonEnableStatePublishSubject,
+    Key? key,
+    required DateTime this.initialDate,
+    required this.onYearSelected,
+    required this.upDownPageLimitPublishSubject,
+    required this.upDownButtonEnableStatePublishSubject,
     this.firstDate,
     this.lastDate,
     this.locale,
@@ -32,7 +32,7 @@ class YearSelector extends StatefulWidget {
 }
 
 class YearSelectorState extends State<YearSelector> {
-  PageController _pageController;
+  PageController? _pageController;
 
   @override
   Widget build(BuildContext context) => PageView.builder(
@@ -56,17 +56,17 @@ class YearSelectorState extends State<YearSelector> {
       );
 
   Widget _getYearButton(final int page, final int index, final String locale) {
-    final int year = (widget.firstDate == null ? 0 : widget.firstDate.year) +
+    final int year = (widget.firstDate == null ? 0 : widget.firstDate!.year) +
         page * 12 +
         index;
     final bool isEnabled = _isEnabled(year);
     return FlatButton(
       onPressed: isEnabled ? () => widget.onYearSelected(year) : null,
-      color: year == widget.initialDate.year
+      color: year == widget.initialDate!.year
           ? Theme.of(context).accentColor
           : null,
-      textColor: year == widget.initialDate.year
-          ? Theme.of(context).accentTextTheme.button.color
+      textColor: year == widget.initialDate!.year
+          ? Theme.of(context).accentTextTheme.button!.color
           : year == DateTime.now().year ? Theme.of(context).accentColor : null,
       child: Text(
         DateFormat.y(locale).format(DateTime(year)),
@@ -78,10 +78,10 @@ class YearSelectorState extends State<YearSelector> {
     widget.upDownPageLimitPublishSubject.add(new UpDownPageLimit(
         widget.firstDate == null
             ? page * 12
-            : widget.firstDate.year + page * 12,
+            : widget.firstDate!.year + page * 12,
         widget.firstDate == null
             ? page * 12 + 11
-            : widget.firstDate.year + page * 12 + 11));
+            : widget.firstDate!.year + page * 12 + 11));
     if (page == 0 || page == _getPageCount() - 1) {
       widget.upDownButtonEnableStatePublishSubject.add(
         new UpDownButtonEnableState(page > 0, page < _getPageCount() - 1),
@@ -91,10 +91,10 @@ class YearSelectorState extends State<YearSelector> {
 
   int _getPageCount() {
     if (widget.firstDate != null && widget.lastDate != null) {
-      if (widget.lastDate.year - widget.firstDate.year <= 12)
+      if (widget.lastDate!.year - widget.firstDate!.year <= 12)
         return 1;
       else
-        return ((widget.lastDate.year - widget.firstDate.year + 1) / 12).ceil();
+        return ((widget.lastDate!.year - widget.firstDate!.year + 1) / 12).ceil();
     } else if (widget.firstDate != null && widget.lastDate == null) {
       return (_getItemCount() / 12).ceil();
     } else if (widget.firstDate == null && widget.lastDate != null) {
@@ -105,11 +105,11 @@ class YearSelectorState extends State<YearSelector> {
 
   int _getItemCount() {
     if (widget.firstDate != null && widget.lastDate != null) {
-      return widget.lastDate.year - widget.firstDate.year + 1;
+      return widget.lastDate!.year - widget.firstDate!.year + 1;
     } else if (widget.firstDate != null && widget.lastDate == null) {
-      return (9999 - widget.firstDate.year);
+      return (9999 - widget.firstDate!.year);
     } else if (widget.firstDate == null && widget.lastDate != null) {
-      return widget.lastDate.year;
+      return widget.lastDate!.year;
     } else
       return 9999;
   }
@@ -118,28 +118,28 @@ class YearSelectorState extends State<YearSelector> {
   void initState() {
     _pageController = new PageController(
         initialPage: widget.firstDate == null
-            ? (widget.initialDate.year / 12).floor()
-            : ((widget.initialDate.year - widget.firstDate.year) / 12).floor());
+            ? (widget.initialDate!.year / 12).floor()
+            : ((widget.initialDate!.year - widget.firstDate!.year) / 12).floor());
     super.initState();
     new Future.delayed(Duration.zero, () {
       widget.upDownPageLimitPublishSubject.add(new UpDownPageLimit(
         widget.firstDate == null
-            ? _pageController.page.toInt() * 12
-            : widget.firstDate.year + _pageController.page.toInt() * 12,
+            ? _pageController!.page!.toInt() * 12
+            : widget.firstDate!.year + _pageController!.page!.toInt() * 12,
         widget.firstDate == null
-            ? _pageController.page.toInt() * 12 + 11
-            : widget.firstDate.year + _pageController.page.toInt() * 12 + 11,
+            ? _pageController!.page!.toInt() * 12 + 11
+            : widget.firstDate!.year + _pageController!.page!.toInt() * 12 + 11,
       ));
       widget.upDownButtonEnableStatePublishSubject.add(
-        new UpDownButtonEnableState(_pageController.page.toInt() > 0,
-            _pageController.page.toInt() < _getPageCount() - 1),
+        new UpDownButtonEnableState(_pageController!.page!.toInt() > 0,
+            _pageController!.page!.toInt() < _getPageCount() - 1),
       );
     });
   }
 
   @override
   void dispose() {
-    _pageController.dispose();
+    _pageController!.dispose();
     super.dispose();
   }
 
@@ -148,32 +148,32 @@ class YearSelectorState extends State<YearSelector> {
       return true;
     else if (widget.firstDate != null &&
         widget.lastDate != null &&
-        year >= widget.firstDate.year &&
-        year <= widget.lastDate.year)
+        year >= widget.firstDate!.year &&
+        year <= widget.lastDate!.year)
       return true;
     else if (widget.firstDate != null &&
         widget.lastDate == null &&
-        year >= widget.firstDate.year)
+        year >= widget.firstDate!.year)
       return true;
     else if (widget.firstDate == null &&
         widget.lastDate != null &&
-        year <= widget.lastDate.year)
+        year <= widget.lastDate!.year)
       return true;
     else
       return false;
   }
 
   void goDown() {
-    _pageController.animateToPage(
-      _pageController.page.toInt() + 1,
+    _pageController!.animateToPage(
+      _pageController!.page!.toInt() + 1,
       duration: const Duration(milliseconds: 400),
       curve: Curves.easeInOut,
     );
   }
 
   void goUp() {
-    _pageController.animateToPage(
-      _pageController.page.toInt() - 1,
+    _pageController!.animateToPage(
+      _pageController!.page!.toInt() - 1,
       duration: const Duration(milliseconds: 400),
       curve: Curves.easeInOut,
     );
